# Fix inefficient label mapping in direct color mode (10-20x speedup)