--- a/napari/layers/labels/labels.py
+++ b/napari/layers/labels/labels.py
@@ -926,10 +926,10 @@ class Labels(_ImageBase):
         ):
             min_label_id = raw_modified.min()
             max_label_id = raw_modified.max()
-            n_unique_labels_upper_bound = max_label_id - min_label_id
+            upper_bound_n_unique_labels = max_label_id - min_label_id
             none_color_index = self._label_color_index[None]
 
-            if n_unique_labels_upper_bound < 1024:
+            if upper_bound_n_unique_labels < 1024:
                 mapping = np.array(
                     [
                         self._label_color_index.get(label_id, none_color_index)