@Czaki 2023-04-13
Our current target for solving this problem is to avoid global remapping of labels on CPU in #3308 but I'm not sure when linked PR will be finished. 

We will try to investigate of the status of #3308 and if the prognosis of the finish will be not optimistic I will be happy to at least partially solve this performance problem by merging this PR. 

If you have time You could also take a look at the linked PR and maybe help to finish it.