@Czaki 2023-04-13
one idea (the changes will be bigger) 

We do not need to know the real set of labels. We only need to have a guarantee that all labels from layers are present in this set (so even if we erase some label, then its present is not a problem). 

So the call of unique could be delayed and done for example 1s after finishing the drawing (with updating the set of every variable used in the drawing). 

It will be more complex and take a long time with deeper dive in napari code, but it also will be useful for solve problems mentioned in #3308 and speedup code even more (as avoid calculate min and max on every refresh).