@jni 2023-04-14
@ksofiyuk How does your benchmark suite fare on #3308? I think that PR makes this one unnecessary, and it's very important (as you'll see in the discussion there, it fixes a *heap* of issues with direct color mode), so I think we should focus on that one...