@ksofiyuk 2023-04-14
@alisterburt thanks for the improving of readability.

1. I updated the threshold for `upper_bound_n_unique_labels` to 65536, as it can handle even that range of labels without almost any performance drop. Actually, we can safely increase it even more, but I think 65536 should cover 99.999% of possible cases.
2. I also updated the benchmark to incorporate a larger range of generated labels (from -10000 to 10000).

In this benchmark with sparse labels from -10000 to 10000, the proposed method shows even much more significant speedup (up to 40x).

The timing of the old algorithm on a new benchmark (provide only the two biggest sizes for clarity):
```
                2048      521±10ms  
                4096     2.63±0.02s 
```

The new algorithm:
```
                2048     17.3±0.2ms 
                4096     62.3±0.8ms 
```

It is 42x speedup!

I tested it on a bit different hardware, so the number doesn't exactly match the results in the first message.