@ksofiyuk 2023-04-14
> To be clear, this is to handle better cases where labels are sparsely distributed, like [0, 1, 12341234, 893475892374]?

@brisvag, no, on the contrary, it fixes the cases when labels are not super sparsely distributed (I'd say it is 99.999% of real world cases). With the last update, you will gain significant speedup from this PR, if your labels can be localized in any continious range which length is less than 65536 elements. Here are some examples, your labels are:
```
[0, 1, 2, 3]   # you will get significant boost here
[-1000, 0, 1, 1000]  # you will also get significant boost here as it falls inside the interval of length 65536
[100000, 100100, 100200, 100300]  # again it falls inside the interval, the new algorithm will be used
[-1000000, 1000000]  # they cannot be localized in a continuous of length 65536, the old method with `np.unique` will be used.
```