@brisvag 2023-04-14
Sorry, with "this" I meant "keeping the old approach for some cases" :P So I'm on board :+1: