@brisvag 2023-04-14
> I actually think the other one is closer than you're implying @brisvag 

Nice, looking forward to it :)