@ksofiyuk 2023-04-14
> Just pushed a commit trying to make sure all the new lines are covered. This should be good to merge if that passes, deepest apologies if it doesn't rofl (it passed locally)

I'm sorry, but this test is not representative for this code :smile:, because the range of labels is going beyond 65536 (2 ** 16), so the previous version of the mapping will be invoked.