@jni 2023-04-14
@ksofiyuk that's the point — the old tests didn't have a branch, but they were "shuttled" to the new code — if you look at the previous codecov comment, it was the *old* code that was no longer covered by tests. Hence the new test.

Anyway, new npe2 release broke our typing action. CC @tlambert03