@ksofiyuk 2023-04-14
> that's the point — the old tests didn't have a branch, but they were "shuttled" to the new code — if you look at the previous codecov comment, it was the _old_ code that was no longer covered by tests. Hence the new test.

I realized it after I posted it, so I deleted my previous comment, but you managed to read it and reply before it :smile: But thanks for the explanation.