@ksofiyuk 2023-04-14
> How does your benchmark suite fare on https://github.com/napari/napari/pull/3308?

@jni I've tested PR #3308 and it indeed has a much more efficient implementation of the label rendering.

This PR:
```
1024    9.17±0.06ms
2048     27.7±0.1ms
4096      103±3ms
8192      409±10ms
16384     1.70±0.02s
```

PR #3308:
```
1024    1.17±0.01ms
2048    4.50±0.06ms
4096     17.5±0.3ms
8192      71.1±8ms
16384      296±9ms
```

So it is around 5-6x faster than the current implementation on large images. However, you can see that even the GPU implementation has exactly the same pattern of growth, it is just faster because of utilizing better hardware. I believe that the label rendering can be significantly improved by some clever algorithmic optimizations which would allow to have a near constant time complexity of rendering on a brush update without the dependency on the resolution of an image.