@Czaki 2023-04-14
> I believe that the label rendering can be significantly improved by some clever algorithmic optimizations which would allow to have a near constant time complexity of rendering on a brush update without the dependency on the resolution of an image.

One of the core problems here is that we transfer the whole layer array to GPU after paint. 
So bigger array means more linear operations. 

I do not have time to investigate how it will be hard to transfer only the updated part of the image. 

But this plus maybe caching some properties should lead to better performance